realease!!!!v1.0

+++ project/Assets/Code/util/RecvStruct.cs
@@ -2,10 +2,10 @@ namespace Code.util
 {
     public class RecvStruct
     {
-        public double Yaw { get; set; }
-        public double Pitch { get; set; }
+        public float Yaw { get; set; }
+        public float Pitch { get; set; }
         public int Shoot { get; set; }
-        public double TimeStamp { get; set; }
+        public float TimeStamp { get; set; }
         public int RequiredImageWidth { get; set; }
         public int RequiredImageHeight { get; set; }
